pyproject.toml
[project]
name = "filing-cabinet"
version = "0.3.31"
description = "A document processing and filing system with OCR capabilities"
authors = [
    {name = "hellisbugfree", email = "fyi_public@protonmail.com"}
]
dependencies = [
    "click>=8.1.3",
    "langdetect>=1.0.9",
    "tabula-py>=2.7.0",
    "pdfplumber>=0.9.0",
    "Pillow>=9.5.0",
    "pytesseract>=0.3.10",
    "PyPDF2>=3.0.0",
    "pdf2image>=1.16.3",
    "python-magic>=0.4.27",
    "typing-extensions>=4.5.0",
    "pathlib>=1.0.1",
    "python-dotenv>=1.0.0",
]

[project.scripts]
filing = "filing_cabinet.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "pytest-cov>=4.1.0",
    "mypy>=1.5.1",
    "isort>=5.12.0",
    "flake8>=6.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["filing_cabinet"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=filing_cabinet"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.isort]
profile = "black"
multi_line_output = 3
